Add project files.

+++ Entities/Concrete/Car.cs
@@ -10,8 +10,8 @@ namespace Entities.Concrete
         public int CarId { get; set; }
         public int BrandId { get; set; } 
         public int ColorId { get; set; }
-        public int ModelYear { get; set; }
-        public double DailyPrice { get; set; }
+        public short ModelYear { get; set; }
+        public decimal DailyPrice { get; set; }
         public string Description { get; set; }
     }
 }
